Add ClamAV manifest; add class to Hiera
This commit adds a manifest (clamav.pp) to Puppet in order to control ClamAV
(http://www.clamav.net). The manifest itself does not directly run ClamAV.
Rather, it contains a cron type to run a shell script at a random number of
minutes (between 0 and 60) past every hour. The shell script's functionality is
covered in da09c66.

This commit also adds the appropriate class instantiated within the Puppet
manifest to Hiera such that this class can be autoloaded using Hiera as a
lightweight ENC.