Adds script to scan /srv/backup-data
This is a script to scan /srv/backup-data, where backed-up data lives. It works
in conjunction with ab42f22 to scan data backed up from source GOV.UK nodes.

Firstly, it echoes the fact that it's going to start running to clamdscan.log,
and then runs clamscan in quiet mode, logging to clamdscan.log, on
/srv/backup-data. Any infected files are moved to /srv/infected.