Add classes to Hiera; rework govuk_backup user
This commit:

- adds the `base::{mounts,user}` classes to Hiera, ready for Hiera to autoload
  them. This repository uses Hiera as a lightweight ENC to facilitate this.

- removes the `govuk_backup` user from the `gds_accounts` module, as having a
  large number of non-standard arguments passed as part of the creation of the
  user was not clean.

- adds `base::user::ssh_key` - the public SSH key for the `govuk_backup` user's
  account.

- adds the `user.pp` manifest, with:
    - an `ssh_key` class parameter for use with Hiera automatic lookup
    - an argument to restrict the shell to `rssh` to only allow rsync
    - an argument to specify the user must belong to it's own group
    - an argument to specify no further groups are to be applied
    - an argument filled with the value of the `ssh_key` Hiera key

When used together, the two arguments to specify which groups the `govuk_backup`
user must belong to restrict the level of permissions applied to the user.
Firstly, they override the default groups for users to belong to by specifying
the `groups` argument with a blank array as a value. Secondly, they override
second groups by specifying that the user must belong to it's own group.