Add SSH class to Hiera
This commit adds the 'attachmentgenie/ssh' module to the list of classes used in
Hiera. We add this here as we are using Hiera as a lightweight ENC, rather than
the more traditional method of creating an `init.pp` file.

The module contains two classes - `ssh::client` and `ssh::server`. Adding
`ssh::server` in to Hiera to begin with - we can add the client if we decide to
change how we run rsync (currently push-based).