Allows rsync in rssh
The rssh Puppet module was added in be7e918, so this commit adds an rssh.pp
manifest in order to configure rssh.

This commit adds an `allow` attribute to the rssh class in order to specify that
only rsync should be allowed. Full details on the values the `allow` attribute
can accept are at https://forge.puppetlabs.com/blom/rssh.