Separate purge_ignore list in to new Hiera YAML
One part of the default `gds_accounts` behaviour is to purge accounts that exist
on the machine but do not exist in the related Hiera array. When testing changes
against this repository in Vagrant, the default user is `hiera`. Puppet will attempt
to clean this user out, and fail as the user is already logged in (as part of
the VM build-out procedure).

We could add this user to the Hiera array, but that would give the user
permissions which is not a sensible thing to do from a security standpoint.
Instead, it is wise to ignore the existence of the user when the machine is
brought up in a development environment.

This commit adds the user to a `purge_ignore` array in Hiera which is only used
when the machine is brought up in Vagrant, as Vagrant identifies itself as a
development environment. Changes to this repository to support this work will
appear in a separate pull request, and will be referenced in this pull request's
comments.