Add logrotate support
The scan-backup-data.sh script logs output to a file in /var/log. By default,
this file will then be allowed to grow indefinitely. We should curtail this by
using logrotate.

This commit adds a logrotate configuration for `/var/log/clamav/clamdscan.log',
which:

- sets the log rotation to daily
- keeps the prior seven days of logs
- compresses stored logs with gzip
- delays compression until next rotation
- doesn't halt on error
- avoids log rotation if the file is empty
- creates a new file after rotation

It also ensures the file is present in the clamav.pp Puppet manifest.