Improve `rssh` configuration
Previously, the `rssh` configuration was being driven by a Puppet manifest.
Testing showed that this didn't work too well, because whilst it was not
possible to get shell access as that user, it was not possible to use rsync as
that user - which it should have been. Investigation showed that it was rssh's
default behaviour to restrict all access to that user unless there is an
rssh.conf configuration file.

This commit adds a configuration file for `rssh`, allowing only rsync, however
keeping the configuration options for other methods included in the file (but
commented out), to make it easy to enable them in future if required. It also
adds the file to the user.pp Puppet manifest so that it is actually created
during a Puppet run.

This tests out fine now.