Amend clamscan program
In it's original form, this script used `clamdscan`, rather than `clamscan`.
During testing, however, it was realised that the script died as `clamdscan` was
unable to access `/srv/infected`.

By default, AppArmor restricts the behaviour of the clamdscan executable, and
investigation using `aa-status` showed that AppArmor was holding `clamdscan` in
enforce mode. I amended AppArmor to put `clamdscan` in to complain mode
(allowing the executable to run, and logging errors, rather than disallowing it
to run entirely). This then allowed `clamdscan` to run successfully.

However, a second issue then arose: `/srv/backup-data/lost+found` was
unaccessible. `clamdscan` doesn't support the --exclude switch, whereas `clamscan`
does. To that end, the command has been changed to support `clamscan`, thereby
allowing us to exclude `lost+found` as we do not particularly care about data
stored within that folder. Unfortunately, `clamdscan` did not like the use of
the `--exclude` switch.

`clamscan` and `clamdscan` have subtle differences, it seems. The former is
meant to be run as one-off and in cron jobs; the latter is daemonized. The
former also processes the virus definitions each time it runs; the latter does
not.

After it was realised that simply changing the command from `clamdscan` to
`clamscan` allowed us to use the `--exclude` switch, I then tested to see if the
AppArmor changes were still required. It turns out that they were not - using
`clamscan` over `clamdscan` was the overall fix to the issue, and the AppArmor
changes were reverted.