Open tcp/5666 in UFW
NRPE listens on tcp/5666 - without this port open, NRPE won't be able to work
correctly. This commit opens the port in UFW from any IP. Opening up this port
might seem like a bit of a security hole - and we can revisit it later if it is
deemed to be - but we will be locking down the hosts which NRPE can interact
with using the NRPE module.

Netcat shows that this commit has opened tcp/5666 (172.16.10.10 is my local
Vagrant VM's IP):

nc -v 172.16.10.10 5666
found 0 associations
found 1 connections:
     1: flags=82<CONNECTED,PREFERRED>
        outif vboxnet7
        src 172.16.10.1 port 62118
        dst 172.16.10.10 port 5666
        rank info not available
        TCP aux info available

Connection to 172.16.10.10 port 5666 [tcp/*] succeeded!