Set NRPE's allowed hosts
NRPE and Nagios interact using a client-server model, with NRPE being the server
and Nagios being the client. Nagios requests data from NRPE, hence why NRPE is
the server. NRPE is, in essence, a way to perform remote code execution so it is
especially important that we restrict which clients can connect to it and
perform code execution - anything that is not our production Nagios instance
should not be allowed to connect to it.

Our Nagios instance will respond, like all machines, from the SNAT IP in the
vSE - 37.26.90.227. It is, therefore, this IP that has been included in the
allowed_hosts directive. The inclusion of 127.0.0.1 is best practice, according
to the web, to allow for testing.