Add generate_fogfile.sh
In order to generate a Fog file to act as a credential store for
vCloud tools when credentials are user-specific and are not stored in any
centralised data store, we need to create a way to dynamically produce a Fog
file.

This commit adds a shell script which truncates the end of `fog_cred.config`
with new, default, values based on the environment variables passed to it by the
user (which are passed to Jenkins at deploy time).