Pin to Puppet 3.4 and Facter 1.7
So that we are consistently using the same version (which is at the time of
writing latest stable and what we use on GOV.UK) instead of picking the
latest from the PuppetLabs repo at provisioning time.

This doesn't address managing the version of Puppet for existing machines.
We probably want to use alphagov/puppet-puppet