Don't write Fog credentials to disk
Read fog credentials directly from environment variables rather than writing
them to disk. This is generally more secure, but also addresses a race
condition (of sorts) whereby the credential file will be left if
`vcloud-launch` were to exit with a non-zero status.

All existing environment variable names have been preserved so that no
changes to the existing Jenkins job is necessary.

Borrowed from gds/pp-pilotis - should be replaced with an authentication
token when [#68989754] provides a generic tool to do so.