Create check_disk NRPE command in Hiera
The pdxcat/nrpe Puppet Forge module allows us to specify allowed NRPE commands.
This ensures that `allowed_hosts` are only allowed to run commands specified by
`allowed_commands`.

We set the arguments to `check_disk` in NRPE to ensure that, for security
reasons, arguments cannot be injected on-the-wire between monitoring host and
remote host.