Remove scan-backup-data.sh
This script removes the scan-backup-data.sh script as we no longer require
it. The back-ups are sufficiently safe given that they're signed with our
GPG key, and ClamAV would not be able to decrypt the tarball without our
private keys (and, for obvious reasons, that ain't going on the box!).