[#70929368] Set resolv.conf to Carrenza's NSs
The `resolvconf` package is currently installed on the machine in Carrenza
but no nameservers are specified so it can't resolve anything.

Fix this by managing resolvconf with a module that we've used elsewhere and
point it at Carrenza's recursive nameservers (obtained by support ticket).

This won't have any effect within Vagrant because it doesn't set nameservers
if it detects that DHCP is being used:

https://github.com/gds-operations/puppet-resolvconf/blob/988f4075fae6e3bcb15e73f4f63aa6a8d3c1ba8c/manifests/init.pp#L30-33