Switch from `-delete` to a `-exec`
This commit changes the `find` command we run on a cron to switch from
`-delete` to `exec rm {} +`. The reasoning for this is that using `-delete`
results in an error:

`find: The -delete action atomatically turns on -depth, but -prune does
nothing when -depth is in effect.  If you want to carry on anyway, just
explicitly use the -depth option`