[#68989754] Use vcloud-login for Jenkins
Cribbed from gds/govuk-provisioning@62a7923d6186c64e23362d44f194027cbc69932b

Use a shell wrapper, similar to that used by the vcloud-* integration tests,
which creates a temporary FOG_RC (not contain the password) and fetches a
`FOG_VCLOUD_TOKEN` using the new `vcloud-login` utility.

This script is sourced rather than executed so that the trap is called when
`jenkins.sh` exits. Rather than when `jenkins_vcloud_login.sh` completes.

I tried to think of ways to make this more re-usable across projects. The
only thing I could think of was rather horrible; a flag to vcloud-login
which outputs a lot of shell script that can be `eval`ed. I think we'll have
to put up with the duplication for now.