Switch from alphagov/puppet-sudo to saz/sudo
This repo was originally using arnouj/sudo but was switched to a git
reference to alphagov/puppet-sudo which was forked from saz/sudo at
version 3.0.1.

There is absolutely no point this repo referring to an outdated fork to
which we have made no changes, so I am changing the Puppetfile to
reference `saz/sudo ~> 3.0.1` in common with our other repos.