Add additional 500gb Assets disk
We're changing assets to have full backups taken by Duplicity weekly, rather
than monthly. For the first few weeks, we'll have a full backup taken under
the old Duplicity policy (ie, a full monthly, rather than a full weekly).

Some maths was done to work out exactly how big we need to make this disc.
Assuming asset-manager has 1gb incremental backups and 6gb full backups, and
incrementals are run six times a week, then this totals 12gb/week of backups
for asset manager. Multiplying this by three (for three weeks, because we
need to keep the existing monthly backup, plus two new backups), is 36gb.

Similarly, for Whitehall assets, an incremental is 1.5gb and a full is
300gb. Incrementals are taken 6 times per week, meaning we need a mininum of
309gb/week. Multiplying this by three (for three weeks, because we need to
keep the existing monthly backup, plus two new backups) is 927gb.

This means that the disk needs to total at least 1.5tb, including some
headway, so this commit adds in the difference between the disc as it stands
and the total size we need in future.