Update list of disk mounts Puppet is aware of
The list of disk mounts that Puppet is aware of is different to those on the
box. Thus, when Puppet is deployed, it tries to remove the disks it's not
aware of. Thankfully, it's not able to do this since the disks are mounted.

Adds /dev/sdg and /dev/sdh to the list of known mounts such that Puppet can
run on deployment.