Allow asset-backup disks to be defined in Hiera
A new off-site backup machine is being provisioned in Skyscape to complement
the existing Carrenza off-site backup machine. The purpose of each disk, as
according to pvdisplay, differs between the Carrenza and Skyscape machines.

Allow the assets-backup disks to be defined in Hiera on a per-node basis,
such that we can control the disks used to create the LVM mounts.

Stop using ::clientcert since its value does not always equal the FQDN.
Instead, use ::fqdn which generates the FQDN using hostname(1)'s -f option.