Allow graphite-backup disks to be defined in Hiera
A new off-site backup machine is being provisioned in Skyscape to complement
the existing Carrenza off-site backup machine. The purpose of each disk, as
according to pvdisplay, differs between the Carrenza and Skyscape machines.

Allow the graphite-backup disks to be defined in Hiera on a per-node basis,
such that we can control the disks used to create the LVM mounts.

Disable the 80char Puppet-Lint check because the require parameter using the
Filesystem resource is over 80 characters long.