Ensure FQDN values are correct
A valid and correct FQDN value is required to ensure that the Hiera
hierarchy can be matched, which is needed to ensure that the
hierarchy-matching YAML file can be found in hieradata/. Without this, the
relevant disks to use on each machine won't be able to be found which breaks
Puppet since each machine needs to know which disks are available to it.
They differ due to
https://github.com/alphagov/govuk_offsitebackups-puppet/pull/105.

- Renames the machine in the vCloud YAML template to ensure that the
  correct value for the FQDN is returned

- Changes the Hiera hierarchy in order to use the ::fqdn certificate rather
  than the ::clientcert, which consistently gives the correct value.
  Enforcing the use of the FQDN is a safer mechanism than using
  ::clientcert, which takes the value of the CN of the PuppetCA signed SSL
  certificate on the master used by a particular agent.

The change to the hierarchy in hiera.yml to replace the use of the
::clientcert fact with the ::fqdn fact is to ensure that we generate a
hierarchy based on the true FQDN, which is populated using the changes
to bootstrap.erb in this file.

Additionally removes whitespace.

Ultimately, this commit allows the fqdn and domain Facter facts to become
present again.