Properly configure DNS and hostname in OS
This snippet properly configures DNS and hostname(1) in the operating system
by adding and amending entries in /etc/hosts, /etc/network/interfaces and
/etc/resolv.conf (using resolvconf(5)). It also renames the machine in the
vCloud Launcher template, as the VM name is used as the hostname by
hostname(1). Renaming it makes it easier to get the correct hostname in to
the FQDN.

- Removes entries relating to dns-search from /etc/network/interfaces
- Removes entries relating to dns-nameservers from /etc/network/interfaces
- Removes 127.0.1.1 (local dnsmasq IP) from /etc/hosts
- Adds correct dns-search entries to /etc/network/interfaces
- Adds correct dns-nameservers entries to /etc/network/interfaces
- Adds correct nameserver and search domain entries to /etc/resolv.conf
- Adds new hierarchy level to hiera.yml to make use of node.{fqdn} datafiles

The change to the hierarchy in hiera.yml to replace the use of the
::clientcert fact with the ::fqdn fact is to ensure that we generate a
hierarchy based on the true FQDN, which is populated using the changes
to bootstrap.erb in this file.

Additionally removes whitespace.

Ultimately, this commit allows the fqdn and domain Facter facts to become
present again.