Variabalise filesystem block mount name
The block mount path used in the ext4mount resource's require parameter is
over eighty characters long, which triggers the Puppet-Lint character count
check. If the character count for the line is equal to 80 or over, it fires
a warning and tests fail.

Variabalise the name, and call the variable in both places it is used
instead to work around this.