Allow asset-backup disks to be defined in Hiera
A new off-site backup machine is being created in Skyscape to complement the
existing Carrenza off-site backup machine. This commit changes the way we
attach physical volumes to logical volumes, by allowing them to be defined
in Hiera, because they differ per machine.

The existing machine in Carrenza has had disks added to it when we've wanted
to add more disk space to the machine. This is inefficient because we'll
eventually hit https://github.com/gds-operations/vcloud-launcher/issues/104
when we provision the eighth disk. Disks before the eighth will provision
fine; disks after will, in theory, provision fine, but this will block them
being created.