Allow graphite disks to be configured in Hiera
A new off-site backup machine is being created in Skyscape to complement the
existing Carrenza off-site backup machine.

This commit changes the way we attach physical volumes to logical volumes by
allowing them to be defined in Hiera, because they differ per machine. The
convention used is as used in 8c9e9dc. The lvm::volume resource doesn't
accept multiple inputs as an array - see
http://projects.puppetlabs.com/issues/22251 - so the only way to get this
working is to unpack the resource. The comment on L70 also replicates
8c9e9dc and makes it clearer what the unpacking of `lvm::volume` is doing.