Rename Carrenza backup node
We use the ::fqdn Facter fact to help us apply the correct node-specific
Hiera to a machine. The value of ::fqdn is incorrect if the name value is
set to offsite-backup-1, since we end up with
`offsite-backup-1.backup.prov..` rather than `backup0.backup.prov...`, which
is the value we need.

Fixes that by changing the value of the name field.