Split out provider setting in to separate tasks
As it stands, the fabfile allowed us to set the environment we wish to run a
task in, however since the hosts chosen would be part of an array, the task
would be run against all hosts.

In order to allow for tasks to be run against specific hosts, separate the
setting of `env.hosts` out in to different tasks which can be chosen at
runtime.

Amends README to clarify this.