Update disks for cdnlogs
In provider0, the mount point `/srv/backup-cdn-logs` has used it's full 60G
worth of disk and will probably continue to rise.

`/srv/backup-logs` has used none of it's 500G and will not be used in
the foreseeable future.

This change seeks to remove the directory and logical volume for
backup-logs and add the disk for use with backup-cdn-logs.