Add another disk to backup0.provider0
This mount is running out of disk space. The job is meant to store two
full backups of a 600GB dataset, but only has 1.5TB of disk space. It
does the cleanup task *after* it has taken the third full backup, so it
actually runs out of space during the backup and never does the cleanup
task.